refactor(view): Rename PlayerRow & move MainMenuView
